requirements.txt
selenium==4.11.2
webdriver-manager==4.0.0
python-dotenv==1.0.0
tk==0.1.0
